Update item validation in Orders model
This commit corrects the item validation in the Orders model. Previously, validation only checked whether the item field was not null. But this leads to problems when the item is an empty string or whitespace. Therefore, the validation has been updated to also require the item value to contain non-whitespace characters.